requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
asynctest==0.13.0
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
discord.py==1.7.3
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.0
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
python-dotenv==0.19.2
requests==2.27.1
tomli==2.0.1
typed-ast==1.5.2
typing-extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
zipp==3.7.0
